Package project with pyproject and reorganize
