+++ examples/run_basic.py
@@ -3,10 +3,10 @@
 import os
 import sys
 
-# Ajoute le répertoire parent pour pouvoir importer ``launcher``
+# Ajoute le répertoire parent pour pouvoir importer le package
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-from launcher import Simulator
+from simulateur_lora_sfrd.launcher import Simulator
 
 if __name__ == "__main__":
     sim = Simulator(num_nodes=20, packet_interval=10, transmission_mode="Random")
