+++ pyproject.toml
@@ -1,0 +1,30 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "simulateur_lora_sfrd"
+version = "0.1.0"
+description = "Simulateur complet de réseau LoRa"
+readme = "README.md"
+requires-python = ">=3.10"
+license = {file = "LICENSE"}
+authors = [{name = "Unknown"}]
+
+
+# Dependencies from requirements.txt
+dependencies = [
+  "numpy>=1.21",
+  "pandas>=1.3",
+  "scipy>=1.7",
+  "matplotlib>=3.5",
+  "plotly>=5.4",
+  "panel>=0.13",
+  "fastapi>=0.88",
+  "uvicorn>=0.20"
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["simulateur_lora_sfrd*"]
+
