+++ simulateur_lora_sfrd/launcher/clean_results.py
