+++ simulateur_lora_sfrd/launcher/tests/test_sensitivity.py
@@ -1,4 +1,4 @@
-from launcher.channel import Channel
+from simulateur_lora_sfrd.launcher.channel import Channel
 
 
 def test_channel_sensitivity_values():
