+++ tests/test_path_mobility.py
@@ -1,6 +1,6 @@
 import json
-from launcher.path_mobility import PathMobility
-from launcher.node import Node
+from simulateur_lora_sfrd.launcher.path_mobility import PathMobility
+from simulateur_lora_sfrd.launcher.node import Node
 
 def test_dynamic_obstacle_blocks_move(tmp_path):
     path_map = [[0, 0], [0, 0]]
